Speed up isAnyRowOpen value change
The values inside rowMap take about 2-3 seconds to update, which means when you close a row by pressing anywhere, no row is clickable for 2-3 seconds.
The onRowDidClose/Open (emphasis on Did) also take 2-3 to raise, so they cannot be used to check for open rows myself
Instead, the onRowClose/Open events were used. Although, these events are raised before the onPress event, which means when a row is closed by pressing elsewhere - first the openRows value change (to null), and only then the onPress runs, which causes it to think no rows were open and raise the onFilePressed event
The temporary way to handle this is to simply add a timeout for the change, just so handleRowPress will run before onRowClose